Add Basic Filtering By Year
Add a `--year` argument for limiting the output data to a given year.
This is a simple filter and does not limit the scope of API requests.
Since we do not have pagination, you only get a full years worth of data
if you have less than 500 total trades. Proper API pagination will fix
this issue at a later point.